+++ index.ts
@@ -1,16 +1,16 @@
 export {
+    isBadRequestException,
     isInvalidParameterException,
     isInvalidSessionException,
     isOccConflictException,
     isResourceNotFoundException,
-    isResourcePreconditionNotMetException
+    isResourcePreconditionNotMetException,
+    isTransactionExpiredException
 } from "./src/errors/Errors";
-export { Executable } from "./src/Executable";
-export { PooledQldbDriver } from "./src/PooledQldbDriver";
 export { QldbDriver } from "./src/QldbDriver";
-export { QldbSession } from "./src/QldbSession";
-export { createQldbWriter, QldbWriter } from "./src/QldbWriter";
 export { Result } from "./src/Result";
 export { Transaction } from "./src/Transaction";
-export { TransactionExecutable } from "./src/TransactionExecutable";
 export { TransactionExecutor } from "./src/TransactionExecutor";
+export { RetryConfig } from "./src/retry/RetryConfig";
+export { BackoffFunction } from "./src/retry/BackoffFunction";
+export { defaultRetryConfig } from "./src/retry/DefaultRetryConfig"