+++ src/retry/BackoffFunction.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
  * the License. A copy of the License is located at
@@ -11,15 +11,4 @@
  * and limitations under the License.
  */
 
-import { makeBinaryWriter, Writer } from "ion-js";
-
-export interface QldbWriter extends Writer { }
-
-/**
- * Create a QLDB writer for parameters.
- * @returns QldbWriter.
- */
-export function createQldbWriter(): QldbWriter {
-    const ionWriter: Writer = makeBinaryWriter();
-    return <QldbWriter> ionWriter;
-}
+export type BackoffFunction =  (retryAttempt: number, error: Error, transactionId: string) => number;